+++ packages/plugin-commons/web/src/Utils/generateInsertPluginButtonProps.ts
@@ -15,9 +15,12 @@ import {
   ToolbarButtonProps,
   Pubsub,
   EditorPluginConfig,
+  GetEditorState,
+  onPluginAddStepArgs,
+  PluginAddParams,
+  SetEditorState,
   Version,
 } from 'wix-rich-content-common';
-import { GetEditorState, onPluginAddStepArgs, SetEditorState } from 'wix-rich-content-common/src';
 import { getPluginParams } from './getPluginParams';
 
 export function generateInsertPluginButtonProps({
@@ -55,11 +58,16 @@ export function generateInsertPluginButtonProps({
   closePluginMenu?: CloseModalFunction;
 }): ToolbarButtonProps {
   const onPluginAdd = () => helpers?.onPluginAdd?.(blockType, toolbarName);
-  const onPluginAddStep = (step: onPluginAddStepArgs['step'], blockKey: string) => {
+  const onPluginAddStep = (
+    step: onPluginAddStepArgs['step'],
+    blockKey: string,
+    params?: PluginAddParams
+  ) => {
     helpers?.onPluginAddStep?.({
       version: Version.currentVersion,
       entryType: toolbarName, //plusButton = SIDE, moreButton = SHORTCUT, footer = FOOTER
       entryPoint: toolbarName,
+      params,
       pluginId: blockType,
       pluginDetails: blockKey,
       step,
@@ -68,18 +76,22 @@ export function generateInsertPluginButtonProps({
   const onPluginAddSuccess = (params = {}) =>
     helpers?.onPluginAddSuccess?.(blockType, toolbarName, params);
 
-  function addBlock(data) {
+  function addBlock(data, beforeAdd?: (blockKey: string, params?: PluginAddParams) => void) {
     const { newBlock, newSelection, newEditorState } = createBlock(
       getEditorState(),
       data,
       blockType
     );
+    const params = getPluginParams(data, blockType);
+    const blockKey = newBlock.getKey();
+    beforeAdd?.(blockKey, params);
     setEditorState(EditorState.forceSelection(newEditorState, newSelection));
-    onPluginAddSuccess(getPluginParams(data, blockType));
+    onPluginAddSuccess(params); //TOOD: support pluginDetails / pluginUniqueId
     return { newBlock, newSelection, newEditorState };
   }
 
   function addCustomBlock(buttonData: InsertButton) {
+    onPluginAdd();
     buttonData.addBlockHandler?.(getEditorState());
     onPluginAddSuccess();
   }
@@ -105,7 +117,6 @@ export function generateInsertPluginButtonProps({
 
   function onClick(event: MouseEvent) {
     event.preventDefault();
-    onPluginAdd();
     switch (button.type) {
       case 'file':
         toggleFileSelection();
@@ -117,6 +128,7 @@ export function generateInsertPluginButtonProps({
         addCustomBlock(button);
         break;
       case BUTTON_TYPES.BUTTON:
+        onPluginAdd();
         if (button.onClick) {
           button.onClick(event);
         } else {
@@ -124,6 +136,7 @@ export function generateInsertPluginButtonProps({
         }
         break;
       default:
+        onPluginAdd();
         addBlock(button.componentData || {});
         break;
     }
@@ -155,6 +168,7 @@ export function generateInsertPluginButtonProps({
   }
 
   function handleExternalFileChanged({ data, error }) {
+    onPluginAdd();
     if (data) {
       const handleFilesAdded = shouldCreateGallery(data)
         ? (blockKey: string) => commonPubsub.getBlockHandler('galleryHandleFilesAdded', blockKey)
@@ -167,6 +181,7 @@ export function generateInsertPluginButtonProps({
   }
 
   function toggleButtonModal(event) {
+    onPluginAdd();
     if (helpers && helpers.openModal) {
       let modalStyles = {};
       if (button.modalStyles) {
@@ -191,9 +206,8 @@ export function generateInsertPluginButtonProps({
         theme,
         componentData: button.componentData,
         onConfirm: obj => {
-          const data = addBlock(obj);
+          const data = addBlock(obj, (key, params) => onPluginAddStep('PluginModal', key, params));
           addedBlockKey = data.newBlock;
-          onPluginAddStep('PluginModal', addedBlockKey);
           return data;
         },
         pubsub,